Apply correctionfactor to SensorData Message
When a message is received, look up the correctionfactor for the temperature reported by this specific sensorId.
Apply the correction factor to the measured temperature and send the modified message to the output endpoint of the module.

When a correction-factor could not be found for the sensorId of the reported message, we'll make sure that a default one is created and that the default correction factor for that sensor is also present in the reported properties of the module.